backend/requirements.txt
fastapi
uvicorn
opencv-python
python-dotenv
torch
transformers
mistralai
pyttsx3
requests
Pillow
python-multipart 
langchain_groq